requirements.txt
pillow==10.2.0
PyYAML==6.0.1
tk==0.1.0
